refactor(operator): extract RegistryTab from SolverNetsSection DISCOVER block; delete SolverNetsSection
Moves RegistryCatalog into RegistryTab (§2.5); SolverNetsSection and
its test are deleted (no longer used). Operator.tsx has SolverNetsSection
removed; routing test updated to assert Data Donation section instead.
